JobsFinder/backend/requirements.txt
flask
flask-cors
openai
requests
beautifulsoup4
selenium
python-dotenv
pymongo
pdfminer.six
python-docx
nltk
scikit-learn
tiktoken
